get rid of process::interface
