--- kernel/src/process/interface/mod.rs
@@ -1,4 +1,0 @@
-mod dispatch_cmd;
-mod owned_queue;
-pub mod resource;
-pub use owned_queue::OwnedQueue;
